ASSIGNMENT 1 REDO UPDATED
Fixed hover states in the navigation
went through css and grouped, and used hex codes for colors.
Included normalize, and header tags
Added lines under the headlines for Features and Works
Changed the color of "I am your tag line"
Did not apply with to text elements.
Included js smooth-scroll.
Added light green dashes to "We are a team of professionals